Fix issue with installation of modules (#320)

+++ omod-1.8/src/main/java/org/openmrs/module/webservices/rest/web/v1_0/resource/openmrs1_8/ModuleActionResource1_8.java
@@ -171,7 +171,7 @@ public class ModuleActionResource1_8 extends BaseDelegatingResource<ModuleAction
 			throw new APIException(e.getMessage(), e);
 		}
 		finally {
-			if (moduleFile != null) {
+			if (tempModule == null && moduleFile != null) {
 				FileUtils.deleteQuietly(moduleFile);
 			}
 		}
